+++ doc/whats-new.rst
@@ -14,7 +14,7 @@ New Features
 ~~~~~~~~~~~~
 - Expose :py:class:`~xarray.indexes.RangeIndex`, and :py:class:`~xarray.indexes.CoordinateTransformIndex` as public api
   under the ``xarray.indexes`` namespace. By `Deepak Cherian <https://github.com/dcherian>`_.
-- New :py:class:`xarray.indexes.TreeIndex`, which by default uses :py:class:`scipy.spatial.KDTree` under the hood for
+- New :py:class:`xarray.indexes.NDPointIndex`, which by default uses :py:class:`scipy.spatial.KDTree` under the hood for
   the selection of irregular, n-dimensional data (:pull:`10478`).
   By `Benoit Bovy <https://github.com/benbovy>`_.